chore(benchmarking): added benchmarks to EndpointTargetsFromServices
